+++ src/Sulu/Bundle/SecurityBundle/Tests/Functional/Command/CreateRoleCommandTest.php
@@ -14,9 +14,8 @@ use Doctrine\Bundle\FixturesBundle\Command\LoadDataFixturesDoctrineCommand;
 use Sulu\Bundle\TestBundle\Testing\SuluTestCase;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Component\Console\Tester\CommandTester;
-use Sulu\Bundle\SecurityBundle\Entity\Role;
 
-class CreateUserCommandTest extends SuluTestCase
+class CreateRoleCommandTest extends SuluTestCase
 {
     /**
      * @var CommandTester
@@ -32,39 +31,21 @@ class CreateUserCommandTest extends SuluTestCase
         $loadFixturesCommandTester = new CommandTester($loadFixturesCommand);
         $loadFixturesCommandTester->execute(array(), array('interactive' => false));
 
-        $createUserCommand = new CreateUserCommand();
+        $createUserCommand = new CreateRoleCommand();
         $createUserCommand->setApplication($application);
         $this->tester = new CommandTester($createUserCommand);
-
-        $doctrine = $this->getContainer()->get('doctrine');
-        $em = $doctrine->getManager();
-        $now = new \Datetime();
-
-        $role = new Role();
-        $role->setName('test');
-        $role->setSystem('Sulu');
-        $role->setCreated($now);
-        $role->setChanged($now);
-
-        $em->persist($role);
-        $em->flush();
     }
 
     public function testExecute()
     {
         $this->tester->execute(
             array(
-                'username' => 'sulu',
-                'firstName' => 'Sulu',
-                'lastName' => 'Hikaru',
-                'email' => 'sulu.hikaru@startrek.com',
-                'locale' => 'en',
-                'role' => 'test',
-                'password' => 'sulu'
+                'name' => 'test',
+                'system' => 'Sulu'
             ),
             array('interactive' => false)
         );
 
-        $this->assertEquals("Created user sulu in role test\n", $this->tester->getDisplay());
+        $this->assertEquals("Created role test in system Sulu\n", $this->tester->getDisplay());
     }
 }